# Dtos/UserRegistrationDto.cs
namespace idleGood_WebAPI.Dtos
{
    public class UserRegistrationDto
    {
        public string Username { get; set; }
        public string Email { get; set; }
        public string Password { get; set; }
    }
}