# Dtos/UserUpdateDto.cs
namespace idleGood_WebAPI.Dtos
{
    public class UserUpdateDto
    {
        public string UserName { get; set; }
        public string Email { get; set; }
        public string Address { get; set; }
        public string SelfIntro { get; set; }
    }
}